refactor : 함수 리팩토링
