+++ src/main/java/racingcar/CarManager.java
@@ -7,6 +7,8 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import camp.nextstep.edu.missionutils.Randoms;
+
 public class CarManager {
 	public List<Car> generate(String carNames) {
 		//TODO 차 입력 예외처리 및 유효성 검사 필요
@@ -33,5 +35,21 @@ public class CarManager {
 			.stream()
 			.map(Car::getPosition)
 			.collect(Collectors.toList()));
+	}
+
+	public List<Car> updateCarListPosition(List<Car> carList) {
+		return carList.stream()
+			.map(this::updateCarRandom)
+			.collect(Collectors.toList());
+	}
+
+	private Car updateCarRandom(Car car) {
+		int randomNumber = Randoms.pickNumberInRange(0, 9);
+
+		if (randomNumber > 4 || randomNumber == 4) {
+			car.updatePosition();
+		}
+
+		return car;
 	}
 }
