+++ src/main/java/racingcar/Game.java
@@ -6,44 +6,32 @@ import java.util.stream.Collectors;
 import camp.nextstep.edu.missionutils.Console;
 import camp.nextstep.edu.missionutils.Randoms;
 import racingcar.constant.Message;
+import racingcar.iomanagement.InputScanner;
 import racingcar.iomanagement.OutputPrinter;
 
 public class Game {
 	CarManager carManager = new CarManager();
+	InputScanner inputScanner = new InputScanner();
 	OutputPrinter outputPrinter = new OutputPrinter();
 
 	public void startGame() {
-		System.out.println(Message.CAR_INPUT);
-		String carNames = Console.readLine();
+		String  carNames = inputScanner.enterCarName();
 		List<Car> carList = carManager.generate(carNames);
-
-		System.out.println(Message.NUMBER_OF_ATTEMPTS);
-		long count = Long.parseLong(Console.readLine());
+		long count = inputScanner.enterCount();
 
 		System.out.println(Message.EXECUTION_RESULT);
-		for (int i = 0; i < count; i++) {
-			carList = updateCarListPosition(carList);
-			carList.forEach(car -> outputPrinter.printStepResult(car));
-			System.out.println();
-		}
+		carList = getResultOfEachAttempt(carList, count);
 
 		List<Car> winnerCarList = carManager.getWinner(carList);
 		outputPrinter.printFinalWinner(winnerCarList);
 	}
 
-	public List<Car> updateCarListPosition(List<Car> carList) {
-		return carList.stream()
-			.map(this::updateCarRandom)
-			.collect(Collectors.toList());
-	}
-
-	private Car updateCarRandom(Car car) {
-		int randomNumber = Randoms.pickNumberInRange(0, 9);
-
-		if (randomNumber > 4 || randomNumber == 4) {
-			car.updatePosition();
+	public List<Car> getResultOfEachAttempt(List<Car> carList, long count) {
+		for (int i = 0; i < count; i++) {
+			carList = carManager.updateCarListPosition(carList);
+			carList.forEach(car -> outputPrinter.printStepResult(car));
+			System.out.println();
 		}
-
-		return car;
+		return carList;
 	}
 }
